Frontend/requirements.txt
pyttsx3==2.6
SpeechRecognition==3.8.1
Keras==2.10.0
Keras_Preprocessing==1.1.2
numpy==1.23.3
scikit-learn==1.1.2
Pillow==9.2.0
stability-sdk==0.3.0
newsapi-python==0.2.6
wolframalpha==5.0.0
wikipedia==1.4.0
speedtest-cli==2.1.3
youtube-search-python==1.6.6
appopener==1.3
WMI==1.5.1
pynput==1.7.6
requests==2.28.1
psutil==5.9.3
python-dotenv==0.21.0
PyAudio==0.2.12
